+++ components/button_test.go
@@ -81,7 +81,11 @@ func TestButton(t *testing.T) {
 	http.ListenAndServe(":8080", nil)
 }
 
-// BenchmarkButtonSubmit-16    	 4754221	       244.8 ns/op	     432 B/op	       3 allocs/op
+// GOEXPERIMENT=nogreenteagc go test -bench=BenchmarkButtonSubmit -benchmem -benchtime=10s
+// BenchmarkButtonSubmit-16        46242216               256.3 ns/op           432 B/op          3 allocs/op
+
+// go test -bench=BenchmarkButtonSubmit -benchmem -benchtime=10s
+// BenchmarkButtonSubmit-16        48235134               247.9 ns/op           432 B/op          3 allocs/op
 func BenchmarkButtonSubmit(b *testing.B) {
 	b.ReportAllocs()
