+++ dsl/attributes_test.go
@@ -19,6 +19,6 @@ func BenchmarkAttrWithValueCall(b *testing.B) {
 	b.ResetTimer()
 
 	for b.Loop() {
-		_ = RenderConvertedHTML(el)
+		_ = RenderFast(el)
 	}
 }
