+++ dsl/class_test.go
@@ -10,6 +10,6 @@ func BenchmarkClassCall(b *testing.B) {
 	b.ResetTimer()
 
 	for b.Loop() {
-		_ = RenderConvertedHTML(el)
+		_ = RenderFast(el)
 	}
 }
