+++ dsl/helpers_if_test.go
@@ -10,6 +10,6 @@ func BenchmarkIfCall(b *testing.B) {
 	b.ResetTimer()
 
 	for b.Loop() {
-		_ = RenderConvertedHTML(el)
+		_ = RenderFast(el)
 	}
 }
