+++ dsl/helpers_text_test.go
@@ -10,6 +10,6 @@ func BenchmarkTextCall(b *testing.B) {
 	b.ResetTimer()
 
 	for b.Loop() {
-		_ = RenderConvertedHTML(el)
+		_ = RenderFast(el)
 	}
 }
