+++ widgets/card/card_horizontal_test.go
@@ -69,7 +69,7 @@ func TestHorizontalCard(t *testing.T) {
 	}
 
 	writer.Write(
-		dsl.RenderConvertedHTML(
+		dsl.RenderFast(
 			page.Build(),
 		),
 	)
