+++ widgets/card/card_vertical_test.go
@@ -56,7 +56,7 @@ func TestVerticalCard(t *testing.T) {
 	}
 
 	writer.Write(
-		dsl.RenderConvertedHTML(
+		dsl.RenderFast(
 			page.Build(),
 		),
 	)
