+++ widgets/card/cards_test.go
@@ -97,7 +97,7 @@ func TestCards(t *testing.T) {
 	}
 
 	writer.Write(
-		dsl.RenderConvertedHTML(
+		dsl.RenderFast(
 			page.Build(),
 		),
 	)
