+++ widgets/header/header_test.go
@@ -52,7 +52,7 @@ func TestHeader(t *testing.T) {
 	}
 
 	writer.Write(
-		dsl.RenderConvertedHTML(
+		dsl.RenderFast(
 			page.Build(),
 		),
 	)
