+++ widgets/hero/hero_test.go
@@ -61,7 +61,7 @@ func TestHero(t *testing.T) {
 	}
 
 	writer.Write(
-		dsl.RenderConvertedHTML(
+		dsl.RenderFast(
 			page.Build(),
 		),
 	)
