+++ widgets/input-date/input_date_test.go
@@ -58,7 +58,7 @@ func TestWidgetInputDate(t *testing.T) {
 	}
 
 	writer.Write(
-		dsl.RenderConvertedHTML(
+		dsl.RenderFast(
 			page.Build(),
 		),
 	)
