+++ widgets/page_test.go
@@ -150,7 +150,7 @@ func TestPage(t *testing.T) {
 	}
 
 	writer.Write(
-		dsl.RenderConvertedHTML(
+		dsl.RenderFast(
 			page.Build(),
 		),
 	)
